清理无用类文件

+++ src/main/java/com/science/gtnl/Utils/machine/EdenGardenManager/EIGMode.java
@@ -20,7 +20,7 @@ public abstract class EIGMode {
 
     public abstract String getName();
 
-    public abstract int getMinVoltageTier();
+    public abstract byte getMinVoltageTier();
 
     public abstract int getMinGlassTier();
